Rework library loading to support conditional unit entry
This makes a handful of changes to the way library dependencies are
processed in order to support conditional unit loading, i.e. not calling
a unit's entry procedure until its code path is actually visited. This
resolves a difference in behaviour between dynamically compiled
libraries whose entrypoints are only called when their code path is
visited, and statically-compiled libraries whose entrypoints are
currently hoisted to the start of the program and always called,
regardless of whether or where they're actually used.

Drop the `file-requirements' hash table in favour of two global lists,
`required-libraries' and `linked-libraries', the first of which is a
superset of the second. The `linked-libraries' list includes everything
that should be linked with the program statically (i.e. as a unit), and
everything else is a runtime dependency (i.e. loaded as a shared object
or source file).

Split the batch driver's `initforms' list into two separate values, one
for import forms (which must be included within the program's wrapper
module's body, if one is used) and one for compiler-introduced
initialisation forms (which must precede the profiling-related
expressions that are inserted into the program when profiling is
enabled, since they're responsible for loading the "profiler" unit).
This is necessary because all libraries now go through the compiler's
logic for `##core#require' nodes, whereas units were previously exempt.
Move all "forms" bindings together in the `let' that introduces them.

Introduce a new `used-libraries' global to keep track of units that are
specified with "-uses" or `(declare (uses))'. These are hoisted to the
top level and called at the start of the program, just like before this
change. The list of `used-units', which is used to generate prototypes
for external unit toplevels in the C backend, is constructed by simply
remembering all `##core#callunit' nodes as they're encounted during
canonicalisation.

Because each import form now has the potential to introduce a call to a
unit entrypoint, `##core#callunit' nodes are deduplicated during CPS.
This ensures that only one call to a given unit's entrypoint will be
generated along a given code path.

Simplify `##sys#process-require' so that it expects just a library,
module, and compilation flag as arguments, and returns just a single
value. Get rid of the global `provided' and `linked-static-extensions'
lists, which are no longer necessary.

For modules that export syntax in static libraries (where module
information is compiled into the libraries themselves), generate code
that will load the module's library dependencies *before* the code for
runtime evaluation of the module's import forms, that is, "(scheme#eval
'(import-syntax ...))". This ensures that static programs do not attempt
to dynamically load the named import libraries dynamically, since
dlopen(3) et al. are disabled by static compilation. We communicate this
situation to `##sys#compiled-module-registration' with a compile mode
flag, for consistency with `##sys#process-require'.

Change the meaning of the "-link" option so that it indicates libraries
that should be linked iff they're required, rather than always
generating a call to their entrypoints (thus requiring them to be linked
unconditionally, as is the case with "-uses"). This option now also
needs to be plumbed through to the "chicken" program, which handles the
differentiation between static and (potentially) dynamic requirements.

Simplify some of the internal procedures in eval.scm. The
`load-library/internal' and `##sys#load-library' procedures can be
combined, as can `load-extension/internal' and `load-extension'. Rename
the internal version of the `load-library' procedure to `load-unit',
since that's really what it does, and use it in the expansion of
`##core#require'.

Drop the `stripu' alias from `process-declaration' and simply call
`strip-syntax' directly instead. `stripu' was only used in two places.

Signed-off-by: felix <felix@call-with-current-continuation.org>