Always write file when -emit-inline-file given
Presently, -emit-inline-file will not emit a file, and will delete one
if it exists, if there are no exported procedures that can be inlined.
This means in a pattern where you always emit for every -unit, and
always consult for every -uses, compilation will halt if a file isn't
found.

It's a matter of opinion what the right thing to do here is, but I'm
inclined to think a file should always be emitted if the flag is
passed, even it's empty, because:

 - It's consistent with how -emit-types-file behaves

 - Usage of the flag indicates user intent to create the file "absence
   of evidence isn't evidence of absence"--the existence of an empty
   file is an affirmative

 - Statement there is nothing to inline

Note the deletion/omission behavor was added on purpose in 2010, so
this isn't accidental, but I'm not sure the reasoning.

Fixes #1715

Signed-off-by: megane <meganeka@gmail.com>
- I copied the rationale from the Trac issue page.
Signed-off-by: felix <felix@call-with-current-continuation.org>