Don't re-use argvector when the CPS call's arguments use rest-ops (#1703)
Code using restops can be compiled to C code essentially like this:

av2 = av;
av2[0]=*((C_word*)lf[0]+1);
av2[1]=t1;
av2[2]=t2;
av2[3]=C_get_rest_arg(c,2,av,2,t0);
tp(4,av2);

But this means the get_rest_arg is taking the newly-written value from
av2 instead of from the original av.

It would be better if we could assign the argument to a temporary
instead so that it precedes the call, but that would require a
restructure of the CPS call itself, which is quite complicated.  For
now this workaround should be fine, as the situation is relatively
rare anyway.

Signed-off-by: felix <felix@call-with-current-continuation.org>